Refactor to support cross-module operation
Previously, the generated Stag class would be directly encoded with the
types it knew about at compilation-time.  When code was generated for
entities in multiple modules this resulted in the Stag class only referencing
classes generated in the outermost module.  In order to retain knowledge
of generated classes across modules and annotation processing passes,
the KnownTypeAdapterFactoriesUtils was added which maintains a list of
all known TypeAdapterFactories in a StagTypeAdapterFactory.list resource,
ServiceLoader style.

Extracted and distributed the parsing logic into each type adapter
implementation, making them fully independent and eliminating a Gson
TypeAdapter cache lookup upon each non-native field access.  In doing
so also fixed a bug where only writer.beginObject was called for null
object values.

Note: It would be trivial to adapt this to a pure ServiceLoader approach
but this approach does not currently work (consistently) on Android
due to toolchain bugs.