fixed bug where generic type adapters could cause stack overflows
If there existed a generic class that contained a reference to itself,
the type adapter that was generated would try to create a new instance
of that type adapter. Because it was creating a new instance of itself,
it caused a stack overflow.

The correct behavior was to recognize that the class was creating a new
version of itself and to reference this instead of instantiating a new
adapter.