Adds tests and hacked solution for use case documented in #155
This code applies a hacky solution that creates a TypeToken
that *appears* to capture the wildcarded generics.  e.g.:

new TypeToken<DynamicallyTypedModel<?>>() {};

This explicitly violates the contract of TypeToken in its Javadoc:

"This syntax cannot be used to create type literals that have wildcard
 parameters, such as Class<?> or List<? extends CharSequence>"

It *appears* to work but is only committed as an intial step towards
finding a better solution.